Migrate existing tests to pytest
